refactor: move reducers into field.js and common.js (1)
