+++ src/game-logic/reducers/helpers.js
@@ -1,0 +1,11 @@
+/**
+ * @param {Array} chancesAndEvents An array of arrays in which the first
+ * element is a number and the second number is a function.
+ * @param {farmhand.state} state
+ * @returns {farmhand.state}
+ */
+export const applyChanceEvent = (chancesAndEvents, state) =>
+  chancesAndEvents.reduce(
+    (acc, [chance, fn]) => (Math.random() <= chance ? fn(acc) : acc),
+    state
+  )
