+++ src/game-logic/reducers/index.js
@@ -5,20 +5,17 @@
 
 import { itemsMap, recipesMap } from '../../data/maps'
 import { levels } from '../../data/levels'
-import { ResourceFactory } from '../../factories'
 import achievements from '../../data/achievements'
 import upgrades from '../../data/upgrades'
 import {
   areHuggingMachinesInInventory,
   canMakeRecipe,
   castToMoney,
-  chooseRandom,
   clampNumber,
   doesInventorySpaceRemain,
   farmProductsSold,
   filterItemIdsToSeeds,
   findCowById,
-  findInField,
   generateCow,
   generateOffspringCow,
   get7DayAverage,
@@ -30,26 +27,18 @@ import {
   getCowMilkItem,
   getCowMilkRate,
   getCowValue,
-  getCropFromItemId,
-  getCropLifeStage,
-  getFinalCropItemIdFromSeedItemId,
-  getInventoryQuantityMap,
   getLevelEntitlements,
-  getPlotContentFromItemId,
   getPlotContentType,
   getPriceEventForCrop,
   getProfit,
   getRandomLevelUpReward,
   getRandomLevelUpRewardQuantity,
   getRandomUnlockedCrop,
-  getRangeCoords,
   getResaleValue,
   getSalePriceMultiplier,
-  getSeedItemIdFromFinalStageCropItemId,
   inventorySpaceRemaining,
   isItemAFarmProduct,
   isItemSoldInShop,
-  isRandomNumberLessThan,
   levelAchieved,
   moneyTotal,
   nullArray,
@@ -64,10 +53,7 @@ import {
   COW_WEIGHT_MULTIPLIER_FEED_BENEFIT,
   COW_WEIGHT_MULTIPLIER_MAXIMUM,
   COW_WEIGHT_MULTIPLIER_MINIMUM,
-  CROW_CHANCE,
   DAILY_FINANCIAL_HISTORY_RECORD_LENGTH,
-  FERTILIZER_BONUS,
-  HOE_LEVEL_TO_SEED_RECLAIM_RATE,
   HUGGING_MACHINE_ITEM_ID,
   LOAN_GARNISHMENT_RATE,
   LOAN_INTEREST_RATE,
@@ -76,30 +62,22 @@ import {
   MAX_LATEST_PEER_MESSAGES,
   MAX_PENDING_PEER_MESSAGES,
   NOTIFICATION_LOG_SIZE,
-  PRECIPITATION_CHANCE,
   PRICE_EVENT_CHANCE,
   PURCHASEABLE_COMBINES,
   PURCHASEABLE_COW_PENS,
   PURCHASEABLE_FIELD_SIZES,
   PURCHASEABLE_SMELTERS,
-  SCARECROW_ITEM_ID,
   SPRINKLER_ITEM_ID,
   STORAGE_EXPANSION_AMOUNT,
-  STORM_CHANCE,
 } from '../../constants'
 import {
   FORGE_AVAILABLE_NOTIFICATION,
-  INVENTORY_FULL_NOTIFICATION,
   OUT_OF_COW_FEED_NOTIFICATION,
-  RAIN_MESSAGE,
-  STORM_MESSAGE,
-  STORM_DESTROYS_SCARECROWS_MESSAGE,
 } from '../../strings'
 import {
   ACHIEVEMENT_COMPLETED,
   COW_ATTRITION_MESSAGE,
   COW_BORN_MESSAGE,
-  CROW_ATTACKED,
   FERTILIZERS_PRODUCED,
   LEVEL_GAINED_NOTIFICATION,
   LOAN_BALANCE_NOTIFICATION,
@@ -112,115 +90,31 @@ import {
   SOLD_ITEM_PEER_NOTIFICATION,
   TOOL_UPGRADED_NOTIFICATION,
 } from '../../templates'
-import {
-  cropLifeStage,
-  fertilizerType,
-  fieldMode,
-  itemType,
-  toolLevel,
-  toolType,
-} from '../../enums'
+import { itemType } from '../../enums'
 
-const { FERTILIZE, OBSERVE, SET_SCARECROW, SET_SPRINKLER } = fieldMode
-const { GROWN } = cropLifeStage
+import { applyChanceEvent } from './helpers'
+
+import {
+  addItemToInventory,
+  decrementItemFromInventory,
+  showNotification,
+} from './common'
+import {
+  applyCrows,
+  modifyFieldPlotAt,
+  processField,
+  processWeather,
+  processSprinklers,
+} from './field'
+
+export * from './common'
+export * from './field'
 
 ///////////////////////////////////////////////////////////
 //
 // Local helper functions
 //
 ///////////////////////////////////////////////////////////
-
-/**
- * @param {?farmhand.crop} crop
- * @returns {?farmhand.crop}
- */
-export const incrementPlotContentAge = crop =>
-  crop && getPlotContentType(crop) === itemType.CROP
-    ? {
-        ...crop,
-        daysOld: crop.daysOld + 1,
-        daysWatered:
-          crop.daysWatered +
-          (crop.wasWateredToday
-            ? 1 +
-              (crop.fertilizerType === fertilizerType.NONE
-                ? 0
-                : FERTILIZER_BONUS)
-            : 0),
-      }
-    : crop
-
-/**
- * @param {?farmhand.plotContent} plotContent
- * @param {boolean} wasWateredToday
- * @returns {?farmhand.plotContent}
- */
-const setWasWateredProperty = (plotContent, wasWateredToday) => {
-  if (plotContent === null) {
-    return null
-  }
-
-  return getPlotContentType(plotContent) === itemType.CROP
-    ? { ...plotContent, wasWateredToday }
-    : { ...plotContent }
-}
-
-/**
- * @param {?farmhand.plotContent} plotContent
- * @returns {?farmhand.plotContent}
- */
-export const setWasWatered = plotContent =>
-  setWasWateredProperty(plotContent, true)
-
-/**
- * @param {?farmhand.plotContent} plotContent
- * @returns {?farmhand.plotContent}
- */
-export const resetWasWatered = plotContent =>
-  setWasWateredProperty(plotContent, false)
-
-/**
- * @param {?farmhand.plotContent} plotContent
- * @returns {?farmhand.plotContent}
- */
-export const resetWasShoveled = plotContent => {
-  if (plotContent && plotContent.isShoveled && plotContent.daysUntilClear > 1) {
-    return {
-      ...plotContent,
-      daysUntilClear: plotContent.daysUntilClear - 1,
-    }
-  }
-
-  return plotContent && !plotContent.isShoveled ? plotContent : null
-}
-
-/**
- * Invokes a function on every plot in a field.
- * @param {Array.<Array.<?farmhand.plotContent>>} field
- * @param {Function(?farmhand.plotContent)} modifierFn
- * @returns {Array.<Array.<?farmhand.plotContent>>}
- */
-const updateField = (field, modifierFn) =>
-  field.map((row, y) => row.map((plot, x) => modifierFn(plot, x, y)))
-
-/**
- * @param {Array} chancesAndEvents An array of arrays in which the first
- * element is a number and the second number is a function.
- * @param {farmhand.state} state
- * @returns {farmhand.state}
- */
-const applyChanceEvent = (chancesAndEvents, state) =>
-  chancesAndEvents.reduce(
-    (acc, [chance, fn]) => (Math.random() <= chance ? fn(acc) : acc),
-    state
-  )
-
-/**
- * @param {farmhand.state} state
- * @returns {farmhand.state}
- */
-export const processWeather = state =>
-  applyChanceEvent([[PRECIPITATION_CHANCE, applyPrecipitation]], state)
 
 /**
  * @param {farmhand.state} state
@@ -310,176 +204,11 @@ const adjustItemValues = state => ({
   ),
 })
 
-/**
- * @param {?farmhand.plotContent} plot
- * @returns {boolean}
- */
-const plotContainsScarecrow = plot => plot && plot.itemId === SCARECROW_ITEM_ID
-
-/**
- * @param {Array.<Array.<?farmhand.plotContent>>} field
- * @returns {boolean}
- */
-const fieldHasScarecrow = field => findInField(field, plotContainsScarecrow)
-
 ///////////////////////////////////////////////////////////
 //
 // Exported reducers
 //
 ///////////////////////////////////////////////////////////
-
-/**
- * @param {farmhand.state} state
- * @returns {farmhand.state}
- */
-export const applyPrecipitation = state => {
-  let { field } = state
-  let scarecrowsConsumedByReplanting = 0
-  let notification
-
-  if (Math.random() < STORM_CHANCE) {
-    if (fieldHasScarecrow(field)) {
-      notification = {
-        message: STORM_DESTROYS_SCARECROWS_MESSAGE,
-        severity: 'error',
-      }
-
-      let { scarecrow: scarecrowsInInventory = 0 } = getInventoryQuantityMap(
-        state.inventory
-      )
-
-      field = updateField(field, plot => {
-        if (!plotContainsScarecrow(plot)) {
-          return plot
-        }
-
-        if (
-          scarecrowsInInventory &&
-          plot.fertilizerType === fertilizerType.RAINBOW
-        ) {
-          scarecrowsInInventory--
-          scarecrowsConsumedByReplanting++
-
-          return plot
-        }
-
-        return null
-      })
-    } else {
-      notification = { message: STORM_MESSAGE, severity: 'info' }
-    }
-  } else {
-    notification = { message: RAIN_MESSAGE, severity: 'info' }
-  }
-
-  state = decrementItemFromInventory(
-    { ...state, field },
-    'scarecrow',
-    scarecrowsConsumedByReplanting
-  )
-
-  state = {
-    ...state,
-    newDayNotifications: [...state.newDayNotifications, notification],
-  }
-
-  state = waterField(state)
-
-  return state
-}
-
-/**
- * @param {farmhand.state} state
- * @returns {farmhand.state}
- */
-export const applyCrows = state => {
-  const { field } = state
-  const newDayNotifications = [...state.newDayNotifications]
-
-  let notificationMessages = []
-
-  const updatedField = fieldHasScarecrow(field)
-    ? field
-    : updateField(field, plotContent => {
-        if (!plotContent || getPlotContentType(plotContent) !== itemType.CROP) {
-          return plotContent
-        }
-
-        const destroyCrop = Math.random() <= CROW_CHANCE
-
-        if (destroyCrop) {
-          notificationMessages.push(
-            CROW_ATTACKED`${itemsMap[plotContent.itemId]}`
-          )
-        }
-
-        return destroyCrop ? null : plotContent
-      })
-
-  if (notificationMessages.length) {
-    newDayNotifications.push({
-      message: notificationMessages.join('\n\n'),
-      severity: 'error',
-    })
-  }
-
-  return { ...state, field: updatedField, newDayNotifications }
-}
-
-/**
- * @param {farmhand.state} state
- * @returns {farmhand.state}
- */
-export const processSprinklers = state => {
-  const { field, itemsSold } = state
-  const crops = new Map()
-  let modifiedField = [...field]
-
-  const { sprinklerRange } = getLevelEntitlements(
-    levelAchieved(farmProductsSold(itemsSold))
-  )
-
-  field.forEach((row, plotY) => {
-    row.forEach((plot, plotX) => {
-      if (!plot || getPlotContentType(plot) !== itemType.SPRINKLER) {
-        return
-      }
-
-      ;[]
-        .concat(
-          // Flatten this 2D array for less iteration below
-          ...getRangeCoords(sprinklerRange, plotX, plotY)
-        )
-        .forEach(({ x, y }) => {
-          const fieldRow = field[y]
-
-          if (!fieldRow) {
-            return
-          }
-
-          const plotContent = fieldRow[x]
-
-          if (
-            plotContent &&
-            getPlotContentType(plotContent) === itemType.CROP
-          ) {
-            if (!crops.has(plotContent)) {
-              modifiedField = modifyFieldPlotAt(
-                { ...state, field: modifiedField },
-                x,
-                y,
-                setWasWatered
-              ).field
-            }
-
-            crops.set(plotContent, { x, y })
-          }
-        })
-    })
-  })
-
-  return { ...state, field: modifiedField }
-}
 
 /**
  * @param {farmhand.state} state
@@ -643,66 +372,6 @@ export const processCowFertilizerProduction = state => {
 }
 
 /**
- * Only adds as many items as there is room in the inventory for unless
- * allowInventoryOverage is true.
- * @param {farmhand.state} state
- * @param {farmhand.item} item
- * @param {number} [howMany=1]
- * @param {boolean} [allowInventoryOverage=false]
- * @returns {farmhand.state}
- */
-export const addItemToInventory = (
-  state,
-  item,
-  howMany = 1,
-  allowInventoryOverage = false
-) => {
-  const { id } = item
-  const inventory = [...state.inventory]
-
-  const numberOfItemsToAdd = allowInventoryOverage
-    ? howMany
-    : Math.min(howMany, inventorySpaceRemaining(state))
-
-  if (numberOfItemsToAdd === 0) {
-    return state
-  }
-
-  const currentItemSlot = inventory.findIndex(({ id: itemId }) => id === itemId)
-
-  if (~currentItemSlot) {
-    const currentItem = inventory[currentItemSlot]
-
-    inventory[currentItemSlot] = {
-      ...currentItem,
-      quantity: currentItem.quantity + numberOfItemsToAdd,
-    }
-  } else {
-    inventory.push({ id, quantity: numberOfItemsToAdd })
-  }
-
-  return { ...state, inventory }
-}
-
-const fieldReducer = (acc, fn) => fn(acc)
-const fieldUpdaters = [
-  incrementPlotContentAge,
-  resetWasWatered,
-  resetWasShoveled,
-]
-
-/**
- * @param {farmhand.state} state
- * @returns {farmhand.state}
- */
-const processField = state => ({
-  ...state,
-  field: updateField(state.field, plotContent =>
-    fieldUpdaters.reduce(fieldReducer, plotContent)
-  ),
-})
-
-/**
  * @param {farmhand.state} state
  * @returns {farmhand.state}
  */
@@ -796,70 +465,6 @@ export const computeCowInventoryForNextDay = state => ({
     isUsingHuggingMachine: Boolean(cow.isUsingHuggingMachine),
   })),
 })
-
-/**
- * @param {farmhand.state} state
- * @returns {farmhand.state}
- */
-export const waterField = state => ({
-  ...state,
-  field: updateField(state.field, setWasWatered),
-})
-
-/**
- * @param {farmhand.state} state
- * @param {number} x
- * @param {number} y
- * @param {Function(?farmhand.plotContent)} modifierFn
- * @returns {farmhand.state}
- */
-export const modifyFieldPlotAt = (state, x, y, modifierFn) => {
-  const { field } = state
-  const row = [...field[y]]
-  const plotContent = modifierFn(row[x])
-  row[x] = plotContent
-  const modifiedField = [...field]
-  modifiedField[y] = row
-
-  return { ...state, field: modifiedField }
-}
-
-/**
- * @param {farmhand.state} state
- * @param {number} x
- * @param {number} y
- * @returns {farmhand.state}
- */
-export const removeFieldPlotAt = (state, x, y) =>
-  modifyFieldPlotAt(state, x, y, () => null)
-
-/**
- * @param {farmhand.state} state
- * @param {string} itemId
- * @param {number} [howMany=1]
- * @returns {farmhand.state}
- */
-export const decrementItemFromInventory = (state, itemId, howMany = 1) => {
-  const inventory = [...state.inventory]
-  const itemInventoryIndex = inventory.findIndex(({ id }) => id === itemId)
-
-  if (itemInventoryIndex === -1) {
-    return state
-  }
-
-  const { quantity } = inventory[itemInventoryIndex]
-
-  if (quantity > howMany) {
-    inventory[itemInventoryIndex] = {
-      ...inventory[itemInventoryIndex],
-      quantity: quantity - howMany,
-    }
-  } else {
-    inventory.splice(itemInventoryIndex, 1)
-  }
-
-  return { ...state, inventory }
-}
 
 /**
  * @param {farmhand.state} state
@@ -1282,42 +887,6 @@ export const upgradeTool = (state, upgrade) => {
   return { ...state }
 }
 
-// TODO: Change showNotification to accept a configuration object instead of so
-// many formal parameters.
-/**
- * @param {farmhand.state} state
- * @param {string} message
- * @param {string} [severity] Corresponds to the `severity` prop here:
- * https://material-ui.com/api/alert/
- * @returns {farmhand.state}
- * @see https://material-ui.com/api/alert/
- */
-export const showNotification = (
-  state,
-  message,
-  severity = 'info',
-  onClick = undefined
-) => {
-  const { showNotifications, todaysNotifications } = state
-
-  return {
-    ...state,
-    ...(showNotifications && {
-      latestNotification: {
-        message,
-        onClick,
-        severity,
-      },
-    }),
-    // Don't show redundant notifications
-    todaysNotifications: todaysNotifications.find(
-      notification => notification.message === message
-    )
-      ? todaysNotifications
-      : todaysNotifications.concat({ message, onClick, severity }),
-  }
-}
-
 /**
  * @param {farmhand.state} state
  * @param {farmhand.cow} cow
@@ -1504,336 +1073,6 @@ export const changeCowBreedingPenResident = (state, cow, doAdd) => {
 
 /**
  * @param {farmhand.state} state
- * @param {number} x
- * @param {number} y
- * @param {string} plantableItemId
- * @returns {farmhand.state}
- */
-export const plantInPlot = (state, x, y, plantableItemId) => {
-  if (
-    !plantableItemId ||
-    !state.inventory.some(({ id }) => id === plantableItemId)
-  ) {
-    return state
-  }
-
-  const { field } = state
-  const row = field[y]
-  const finalCropItemId = getFinalCropItemIdFromSeedItemId(plantableItemId)
-
-  if (row[x]) {
-    // Something is already planted in field[x][y]
-    return state
-  }
-
-  state = modifyFieldPlotAt(state, x, y, () =>
-    getCropFromItemId(finalCropItemId)
-  )
-
-  state = decrementItemFromInventory(state, plantableItemId)
-  state = processSprinklers(state)
-
-  return {
-    ...state,
-    selectedItemId: state.inventory.find(({ id }) => id === plantableItemId)
-      ? plantableItemId
-      : '',
-  }
-}
-
-const fertilizerItemIdToTypeMap = {
-  [itemsMap['fertilizer'].id]: fertilizerType.STANDARD,
-  [itemsMap['rainbow-fertilizer'].id]: fertilizerType.RAINBOW,
-}
-
-/**
- * Assumes that state.selectedItemId references an item with type ===
- * itemType.FERTILIZER.
- * @param {farmhand.state} state
- * @param {number} x
- * @param {number} y
- * @returns {farmhand.state}
- */
-export const fertilizePlot = (state, x, y) => {
-  const { field, selectedItemId } = state
-  const row = field[y]
-  const plotContent = row[x]
-
-  if (!plotContent || itemsMap[selectedItemId]?.type !== itemType.FERTILIZER) {
-    return state
-  }
-
-  const { id: fertilizerItemId } = itemsMap[selectedItemId]
-
-  const fertilizerInventory = state.inventory.find(
-    item => item.id === fertilizerItemId
-  )
-
-  const plotContentType = getPlotContentType(plotContent)
-
-  if (
-    !plotContent ||
-    !fertilizerInventory ||
-    plotContent.fertilizerType !== fertilizerType.NONE ||
-    (selectedItemId === 'fertilizer' && plotContentType !== itemType.CROP) ||
-    (selectedItemId === 'rainbow-fertilizer' &&
-      plotContentType !== itemType.CROP &&
-      plotContentType !== itemType.SCARECROW)
-  ) {
-    return state
-  }
-
-  const { quantity: initialFertilizerQuantity } = fertilizerInventory
-  state = decrementItemFromInventory(state, fertilizerItemId)
-  const doFertilizersRemain = initialFertilizerQuantity > 1
-
-  state = modifyFieldPlotAt(state, x, y, crop => ({
-    ...crop,
-    fertilizerType: fertilizerItemIdToTypeMap[fertilizerItemId],
-  }))
-
-  return {
-    ...state,
-    fieldMode: doFertilizersRemain ? FERTILIZE : OBSERVE,
-    selectedItemId: doFertilizersRemain ? fertilizerItemId : '',
-  }
-}
-
-/**
- * @param {farmhand.state} state
- * @param {number} x
- * @param {number} y
- * @returns {farmhand.state}
- */
-export const setSprinkler = (state, x, y) => {
-  const { field } = state
-  const plot = field[y][x]
-
-  // Only set sprinklers in empty plots
-  if (plot !== null) {
-    return state
-  }
-
-  state = decrementItemFromInventory(state, SPRINKLER_ITEM_ID)
-
-  const doSprinklersRemain = state.inventory.some(
-    item => item.id === SPRINKLER_ITEM_ID
-  )
-
-  state = modifyFieldPlotAt(state, x, y, () =>
-    getPlotContentFromItemId(SPRINKLER_ITEM_ID)
-  )
-
-  state = processSprinklers(state)
-
-  return {
-    ...state,
-    fieldMode: doSprinklersRemain ? SET_SPRINKLER : OBSERVE,
-    selectedItemId: doSprinklersRemain ? SPRINKLER_ITEM_ID : '',
-  }
-}
-
-/**
- * @param {farmhand.state} state
- * @param {number} x
- * @param {number} y
- * @returns {farmhand.state}
- */
-export const setScarecrow = (state, x, y) => {
-  const plot = state.field[y][x]
-
-  // Only set scarecrows in empty plots
-  if (plot !== null) {
-    return state
-  }
-
-  state = decrementItemFromInventory(state, SCARECROW_ITEM_ID)
-
-  const doScarecrowsRemain = state.inventory.some(
-    item => item.id === SCARECROW_ITEM_ID
-  )
-
-  state = modifyFieldPlotAt(state, x, y, () =>
-    getPlotContentFromItemId(SCARECROW_ITEM_ID)
-  )
-
-  return {
-    ...state,
-    fieldMode: doScarecrowsRemain ? SET_SCARECROW : OBSERVE,
-    selectedItemId: doScarecrowsRemain ? SCARECROW_ITEM_ID : '',
-  }
-}
-
-/**
- * @param {farmhand.state} state
- * @param {number} x
- * @param {number} y
- * @returns {farmhand.state}
- */
-export const harvestPlot = (state, x, y) => {
-  const row = state.field[y]
-  const crop = row[x]
-
-  if (
-    !crop ||
-    getPlotContentType(crop) !== itemType.CROP ||
-    getCropLifeStage(crop) !== GROWN ||
-    !doesInventorySpaceRemain(state)
-  ) {
-    return state
-  }
-
-  const item = itemsMap[crop.itemId]
-  const seedItemIdForCrop = getSeedItemIdFromFinalStageCropItemId(item.id)
-  const plotWasRainbowFertilized =
-    crop.fertilizerType === fertilizerType.RAINBOW
-
-  let harvestedQuantity = 1
-
-  switch (state.toolLevels[toolType.SCYTHE]) {
-    case toolLevel.BRONZE:
-      harvestedQuantity += 1
-      break
-
-    case toolLevel.IRON:
-      harvestedQuantity += 2
-      break
-
-    case toolLevel.SILVER:
-      harvestedQuantity += 3
-      break
-
-    case toolLevel.GOLD:
-      harvestedQuantity += 4
-      break
-
-    default:
-      harvestedQuantity = 1
-  }
-
-  state = removeFieldPlotAt(state, x, y)
-  state = addItemToInventory(state, item, harvestedQuantity)
-
-  const { cropType } = item
-
-  if (plotWasRainbowFertilized) {
-    const seedsForHarvestedCropAreAvailable =
-      getInventoryQuantityMap(state.inventory)[seedItemIdForCrop] > 0
-
-    if (seedsForHarvestedCropAreAvailable) {
-      state = plantInPlot(state, x, y, seedItemIdForCrop)
-      state = modifyFieldPlotAt(state, x, y, crop => ({
-        ...crop,
-        fertilizerType: fertilizerType.RAINBOW,
-      }))
-    }
-  }
-
-  const { cropsHarvested } = state
-
-  return {
-    ...state,
-    cropsHarvested: {
-      ...cropsHarvested,
-      [cropType]: (cropsHarvested[cropType] || 0) + harvestedQuantity,
-    },
-  }
-}
-
-/**
- * @param {farmhand.state} state
- * @param {number} x
- * @param {number} y
- * @returns {farmhand.state}
- */
-export const minePlot = (state, x, y) => {
-  const { field } = state
-  const row = field[y]
-
-  if (row[x]) {
-    // Something is already planted in field[x][y]
-    return state
-  }
-
-  if (!doesInventorySpaceRemain(state)) {
-    return showNotification(state, INVENTORY_FULL_NOTIFICATION)
-  }
-
-  const shovelLevel = state.toolLevels[toolType.SHOVEL]
-  const spawnedResources = ResourceFactory.instance().generateResources(
-    shovelLevel
-  )
-  let spawnedOre = null
-  let daysUntilClear = chooseRandom([1, 2, 2, 3])
-
-  if (spawnedResources.length) {
-    // even when multiple resources are spawned, the first one is ok to use
-    // for all subsequent logic
-    spawnedOre = spawnedResources[0]
-
-    // if ore was spawned, add up to 10 days to the time to clear
-    // at random, based loosely on the spawnChance meant to make
-    // rarer ores take longer to cooldown
-    daysUntilClear += Math.round(
-      Math.random() * (1 - spawnedOre.spawnChance) * 10
-    )
-
-    for (let resource of spawnedResources) {
-      state = addItemToInventory(state, resource)
-    }
-  }
-
-  state = modifyFieldPlotAt(state, x, y, () => {
-    return {
-      isShoveled: true,
-      daysUntilClear,
-      oreId: spawnedOre ? spawnedOre.id : null,
-    }
-  })
-
-  return {
-    ...state,
-  }
-}
-
-/**
- * @param {farmhand.state} state
- * @param {number} x
- * @param {number} y
- * @returns {farmhand.state}
- */
-export const clearPlot = (state, x, y) => {
-  const plotContent = state.field[y][x]
-  const hoeLevel = state.toolLevels[toolType.HOE]
-
-  if (!plotContent || plotContent.isShoveled) {
-    return state
-  }
-
-  if (
-    getPlotContentType(plotContent) === itemType.CROP &&
-    getCropLifeStage(plotContent) !== GROWN &&
-    isRandomNumberLessThan(HOE_LEVEL_TO_SEED_RECLAIM_RATE[hoeLevel])
-  ) {
-    const seedId = getSeedItemIdFromFinalStageCropItemId(plotContent.itemId)
-    state = addItemToInventory(state, itemsMap[seedId])
-  }
-
-  const item = itemsMap[plotContent.itemId]
-
-  if (item.isReplantable && !doesInventorySpaceRemain(state)) {
-    return state
-  }
-
-  state = removeFieldPlotAt(state, x, y)
-
-  return item.isReplantable || getCropLifeStage(plotContent) === GROWN
-    ? addItemToInventory(state, item)
-    : state
-}
-
-/**
- * @param {farmhand.state} state
  * @param {number} fieldId
  * @returns {farmhand.state}
  */
@@ -1874,12 +1113,6 @@ export const waterPlot = (state, x, y) => {
     wasWateredToday: true,
   }))
 }
-
-/**
- * @param {farmhand.state} state
- * @returns {farmhand.state}
- */
-export const waterAllPlots = state => waterField(state)
 
 /**
  * @param {farmhand.state} state
